+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added:
 - Tests for measurement in journal
 - Tests for health entry in journal
-- Tests for dashboards measurement chart
+- Tests for dashboards measurement charts
+- Tests for dashboards health charts
+- Tests for dashboards workout charts
 
 ## [0.8.95] - 2022-07-07
 ### Changed: